.history/backend/requirements_20250902215229.txt
fastapi==0.63.0
uvicorn==0.13.4
pymongo==3.11.4
python-dotenv==0.15.0
bcrypt==3.2.0
stripe==2.55.0
python-jose==3.2.0